almost working cart view

+++ src/main/java/dat/startcode/model/DTO/DTOShoppingCart.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 public class DTOShoppingCart{
 
     @Getter private ArrayList<DTOOrderLine> orderLines = new ArrayList<>();
-    @Getter private int total = 0;
+    @Getter private float total = 0;
 
     public void sum() {
         for (DTOOrderLine i : orderLines) {
